backend/pyproject.toml
[tool.poetry]
name = "coax-meet"
version = "0.1.0"
description = ""
authors = ["v-bazhenov <vitalii.bazhenov@coaxsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.2.2"
channels = "^4.0.0"
psycopg2-binary = "^2.9.6"
daphne = "^4.0.0"
channels-redis = "^4.1.0"
drf-yasg = "^1.21.5"
environs = "^9.5.0"
djoser = "^2.2.0"
celery = {extras = ["sqs"], version = "^5.2.7"}
django-celery-email = "^3.0.0"
django-celery-beat = "^2.5.0"
ipython = "^8.13.2"
drf-extra-fields = "^3.4.1"
flower = "^1.2.0"
django-filter = "^23.2"
sendgrid = "^6.10.0"
django-sendgrid-v5 = "^1.2.2"
mkdocs = "^1.4.3"
mkdocs-awesome-pages-plugin = "^2.9.1"
mkdocs-material = "^9.1.15"
boto3 = "^1.26.142"
werkzeug = "^2.3.4"
apig-wsgi = "^2.16.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
coverage = "^7.2.5"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[requires]
python_version = "3.11"
